+++ tests/tests.cpp
@@ -59,16 +59,16 @@ void lakes()
 {
     // Load image
     CImg<double> img = rgb2gray(CImg<double>("../testimages/lakes3.jpg")) / 255.0;
-    
+
     // Initialize embedding function
     auto embedding = circle_levelset(img.height(), img.width(), {80, 170}, 25);
     (embedding * 255).save_png("lakes_begin.png");
-    
+
     // Morphological ACWE
     ms::MorphACWE<double, 2> macwe(cimg2ndimage(embedding), cimg2ndimage(img), 3);
     for(int i = 0; i < 200; ++i)
         macwe.step();
-    
+
     // Save results
     (embedding * 255).save_png("lakes_end.png");
 }
@@ -77,7 +77,7 @@ void starfish()
 {
     // Load image
     CImg<double> img = rgb2gray(CImg<double>("../testimages/seastar2.png")) / 255.0;
-    
+
     // Initialize embedding function
     auto embedding = circle_levelset(img.height(), img.width(), {163, 137}, 135);
     (embedding * 255).save_png("seastar_begin.png");
